worked on memory setter/getter

+++ src/DataMemory.h
@@ -9,17 +9,17 @@
 class DataMemory
 {
 public:
-    DataMemory(uint64_t _size, uint64_t _offset);
+    DataMemory(uint32_t _size, uint32_t _offset);
     virtual ~DataMemory();
-    uint64_t GetSize();
+    uint32_t GetSize();
 
-    uint8_t Get(uint16_t address);
-    void Set(uint16_t address, uint8_t value);
+    uint8_t Get(uint32_t address);
+    void Set(uint32_t address, uint8_t value);
     uint8_t *GetDataPtr();
 
 private:
-    uint64_t size;
-    uint64_t offset;
+    uint32_t size;
+    uint32_t offset;
     uint8_t* data;
 };
